+++ Src/Library/Auth/IRefreshTokenService.cs
@@ -3,5 +3,5 @@
 [HideFromDocs]
 public interface IRefreshTokenService<TResponse>
 {
-    internal Task<TResponse> CreateToken(string userId, Action<UserPrivileges> privileges);
+    internal Task<TResponse> CreateToken(string userId, Action<UserPrivileges>? privileges, object? request);
 }